requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-crispy-forms==2.3
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.10
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
requests==2.32.3
shiboken6==6.8.1
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
